make validator function part of PoochSomething (tmpname) class
